reto09/requirements.txt
Click==7.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
PyMySQL==0.9.3
python-dotenv==0.10.3
Werkzeug==0.16.0
